requirements.txt
fastapi
uvicorn
sqlalchemy
jinja2
python-multipart
pydantic
typing-extensions
python-dotenv